result/[id]/page.tsxリファクタリング
